+++ adapters/browsers/playwright/page.go
@@ -1,5 +1,3 @@
-//go:build !rod
-
 package playwright
 
 import (